Updates clap to v2
This commit updates `clap` to v2.4.0 which contains a bunch of bug fixes
and improvements. This PR also includes building `clap` with all default
features as the old `color` feature is no longer built on Windows
systems and therefore no longer an issue.

Besides the bump in version of `clap`, and associated simple name changes (of
which only 3 affected this crate) such as `add_all`->`args` (etc.), the
`owned_vec_string` function has been changed slightly to reflect `clap`s returning
an `Iterator<Item=&str>` instead of a `Vec<&str>`.